CFY-6089 Resource manager and REST module refactor - Move all rest responses to same file (responses.py) - Move the single swagger function under the endpoint mapper module
