+++ rest-service/manager_rest/rest/endpoint_mapper.py
@@ -12,7 +12,10 @@
 #  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  * See the License for the specific language governing permissions and
 #  * limitations under the License.
-from . import swagger, resources_v1, resources_v2, resources_v2_1, resources_v3
+
+from flask_restful_swagger import swagger
+
+from . import resources_v1, resources_v2, resources_v2_1, resources_v3
 
 SUPPORTED_API_VERSIONS = [('v1', resources_v1),
                           ('v2', resources_v2),
@@ -83,6 +86,43 @@ def setup_resources(api):
         _set_versioned_urls(api, resource, endpoint_suffix)
 
 
+def _add_swagger_resource(api, api_version, resource, resource_path):
+    """
+    This method is based on swagger's
+    :func:'flask_restful_swagger.swagger.docs.add_resource' and modifies it to
+    support multiple API versions. These changes were made:
+
+    1. This method should be called directly for every API resource created.
+    AVOID calling :func:'flask_restful_swagger.swagger.docs', or it will
+    override Flask-Restful's 'add_resource' with swagger's implementation!
+        e.g. add_swagger_resource(api, api_version, resource, '/v1/endpoint')
+
+    2. This method only registers swagger APIs. A separate call to
+    Flask-Restful's 'add_resource' is required in order to register the "real"
+    API resource.
+        e.g: api.add_resource(resource, endpoint_url)
+
+    3. swagger's endpoint_paths contain the api version now, to support
+    multiple versions of the same resource on the same API
+    """
+    endpoint = swagger.swagger_endpoint(resource, resource_path)
+    # Add a .help.json help url
+    swagger_path = swagger.extract_swagger_path(resource_path)
+    endpoint_path = "{0}_{1}_help_json".format(api_version, resource.__name__)
+    api.add_resource(endpoint, "%s.help.json" % swagger_path,
+                     endpoint=endpoint_path)
+    # Add a .help.html help url
+    endpoint_path = "{0}_{1}_help_html".format(api_version, resource.__name__)
+    api.add_resource(endpoint, "%s.help.html" % swagger_path,
+                     endpoint=endpoint_path)
+    swagger.register_once(
+        add_resource_func=api.add_resource, apiVersion=api_version,
+        swaggerVersion='1.2',
+        basePath='http://localhost:8100',
+        resourcePath='/', produces=["application/json"],
+        endpoint='/api/spec')
+
+
 def _set_versioned_urls(api, resource_name, endpoint_suffix):
     resource = None
     for version in SUPPORTED_API_VERSIONS:
@@ -95,4 +135,4 @@ def _set_versioned_urls(api, resource_name, endpoint_suffix):
             endpoint = '{0}/{1}'.format(version_name, endpoint_suffix)
             url = '/api/{0}'.format(endpoint)
             api.add_resource(resource, url, endpoint=endpoint)
-            swagger.add_swagger_resource(api, version_name, resource, url)
+            _add_swagger_resource(api, version_name, resource, url)
